Avoid bundling node:path into client helpers
