+++ src/lib/dashboard/analytics/reviews.ts
@@ -1,8 +1,8 @@
 import {
   averageBusinessResponseHours,
   DEPENDABOT_FILTER,
-  HOLIDAY_SET,
 } from "@/lib/dashboard/analytics/shared";
+import { HOLIDAY_SET } from "@/lib/dashboard/business-days";
 import type { HeatmapCell } from "@/lib/dashboard/types";
 import { query } from "@/lib/db/client";
